+++ src/Detector/Strategy/SuffixTree/Sentinel.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Systemsdk\PhpCPD\Detector\Strategy\SuffixTree;
 
-use function random_int;
-
 /**
  * A sentinel character which can be used to produce explicit leaves for all
  * suffixes. The sentinel just has to be appended to the list before handing
@@ -14,14 +12,8 @@ use function random_int;
  */
 class Sentinel extends AbstractToken
 {
-    /**
-     * @var int The hash value used.
-     */
-    private int $hash;
-
     public function __construct()
     {
-        $this->hash = random_int(0, PHP_INT_MAX);
         $this->tokenCode = -1;
         $this->line = -1;
         $this->file = '<no file>';
@@ -32,11 +24,6 @@ class Sentinel extends AbstractToken
     public function __toString(): string
     {
         return '$';
-    }
-
-    public function hashCode(): int
-    {
-        return $this->hash;
     }
 
     public function equals(AbstractToken $other): bool