+++ src/Log/PMD.php
@@ -6,15 +6,11 @@ namespace Systemsdk\PhpCPD\Log;
 
 use DateTime;
 use DateTimeInterface;
-use DOMElement;
+use Dom\Element;
 use DOMException;
-use DOMNode;
 use Systemsdk\PhpCPD\Cli\Application;
 use Systemsdk\PhpCPD\CodeCloneMap;
 use Systemsdk\PhpCPD\Exceptions\LoggerException;
-
-use function is_object;
-use function sprintf;
 
 final class PMD extends AbstractXmlLogger
 {
@@ -40,9 +36,6 @@ final class PMD extends AbstractXmlLogger
     private const string ATTRIBUTE_PATH_NAME = 'path';
     private const string ATTRIBUTE_LINE_NAME = 'line';
     private const string ATTRIBUTE_END_LINE_NAME = 'endline';
-    private const string ACTION_CREATE_ELEMENT = 'create-element';
-    private const string ACTION_APPEND_CHILD = 'append-child';
-    private const string ACTION_SET_ATTRIBUTE = 'set-attribute';
 
     /**
      * @throws LoggerException
@@ -50,114 +43,50 @@ final class PMD extends AbstractXmlLogger
     public function processClones(CodeCloneMap $clones): void
     {
         try {
+            /** @var Element $cpd */
             $cpd = $this->document->createElement(self::CPD_ELEMENT_NAME);
-            $this->checkResult($cpd, self::CPD_ELEMENT_NAME, self::ACTION_CREATE_ELEMENT);
+            $this->document->appendChild($cpd);
 
-            $result = $this->document->appendChild($cpd);
-            $this->checkResult($result, self::CPD_ELEMENT_NAME, self::ACTION_APPEND_CHILD);
-
-            /** @var DOMElement $result */
-            $result->setAttribute(self::ATTRIBUTE_XMLNS_NAME, self::ATTRIBUTE_XMLNS_VALUE);
-
-            $resultAttr = $result->setAttribute(self::ATTRIBUTE_XMLNS_XSI_NAME, self::ATTRIBUTE_XMLNS_XSI_VALUE);
-            $this->checkResult($resultAttr, self::ATTRIBUTE_XMLNS_XSI_NAME, self::ACTION_SET_ATTRIBUTE);
-
-            $resultAttr = $result->setAttribute(self::ATTRIBUTE_PHPCPD_VERSION_NAME, Application::VERSION);
-            $this->checkResult($resultAttr, self::ATTRIBUTE_PHPCPD_VERSION_NAME, self::ACTION_SET_ATTRIBUTE);
-
-            $resultAttr = $result->setAttribute(
+            $cpd->setAttribute(self::ATTRIBUTE_XMLNS_NAME, self::ATTRIBUTE_XMLNS_VALUE);
+            $cpd->setAttribute(self::ATTRIBUTE_XMLNS_XSI_NAME, self::ATTRIBUTE_XMLNS_XSI_VALUE);
+            $cpd->setAttribute(self::ATTRIBUTE_PHPCPD_VERSION_NAME, Application::VERSION);
+            $cpd->setAttribute(
                 self::ATTRIBUTE_TIMESTAMP_NAME,
-                (new DateTime())->format(DateTimeInterface::ATOM)
+                new DateTime()->format(DateTimeInterface::ATOM)
             );
-            $this->checkResult($resultAttr, self::ATTRIBUTE_TIMESTAMP_NAME, self::ACTION_SET_ATTRIBUTE);
-
-            $resultAttr = $result->setAttribute(self::ATTRIBUTE_VERSION_NAME, self::ATTRIBUTE_VERSION_1_VALUE);
-            $this->checkResult($resultAttr, self::ATTRIBUTE_VERSION_NAME, self::ACTION_SET_ATTRIBUTE);
-
-            $resultAttr = $result->setAttribute(
+            $cpd->setAttribute(self::ATTRIBUTE_VERSION_NAME, self::ATTRIBUTE_VERSION_1_VALUE);
+            $cpd->setAttribute(
                 self::ATTRIBUTE_XSI_SCHEMA_LOCATION_NAME,
                 self::ATTRIBUTE_XSI_SCHEMA_LOCATION_VALUE
             );
-            $this->checkResult($resultAttr, self::ATTRIBUTE_XSI_SCHEMA_LOCATION_NAME, self::ACTION_SET_ATTRIBUTE);
 
             foreach ($clones as $clone) {
-                $duplicationEl = $this->document->createElement(self::DUPLICATION_ELEMENT_NAME);
-                $this->checkResult($duplicationEl, self::DUPLICATION_ELEMENT_NAME, self::ACTION_CREATE_ELEMENT);
+                /** @var Element $duplication */
+                $duplication = $this->document->createElement(self::DUPLICATION_ELEMENT_NAME);
+                $cpd->appendChild($duplication);
 
-                $duplication = $cpd->appendChild($duplicationEl);
-                $this->checkResult($duplication, self::DUPLICATION_ELEMENT_NAME, self::ACTION_APPEND_CHILD);
-
-                /** @var DOMElement $duplication */
-                $result1 = $duplication->setAttribute(self::ATTRIBUTE_LINES_NAME, (string)$clone->numberOfLines());
-                $this->checkResult($result1, self::ATTRIBUTE_LINES_NAME, self::ACTION_SET_ATTRIBUTE);
-
-                $result2 = $duplication->setAttribute(self::ATTRIBUTE_TOKENS_NAME, (string)$clone->numberOfTokens());
-                $this->checkResult($result2, self::ATTRIBUTE_TOKENS_NAME, self::ACTION_SET_ATTRIBUTE);
+                $duplication->setAttribute(self::ATTRIBUTE_LINES_NAME, (string)$clone->numberOfLines());
+                $duplication->setAttribute(self::ATTRIBUTE_TOKENS_NAME, (string)$clone->numberOfTokens());
 
                 foreach ($clone->files() as $codeCloneFile) {
-                    $fileEl = $this->document->createElement(self::FILE_ELEMENT_NAME);
-                    $this->checkResult($fileEl, self::FILE_ELEMENT_NAME, self::ACTION_CREATE_ELEMENT);
+                    /** @var Element $file */
+                    $file = $this->document->createElement(self::FILE_ELEMENT_NAME);
+                    $duplication->appendChild($file);
 
-                    $file = $duplication->appendChild($fileEl);
-                    $this->checkResult($file, self::FILE_ELEMENT_NAME, self::ACTION_APPEND_CHILD);
-
-                    /** @var DOMElement $file */
-                    $result1 = $file->setAttribute(self::ATTRIBUTE_LINE_NAME, (string)$codeCloneFile->startLine());
-                    $this->checkResult($result1, self::ATTRIBUTE_LINE_NAME, self::ACTION_SET_ATTRIBUTE);
-
-                    $result2 = $file->setAttribute(self::ATTRIBUTE_END_LINE_NAME, (string)$codeCloneFile->endLine());
-                    $this->checkResult($result2, self::ATTRIBUTE_END_LINE_NAME, self::ACTION_SET_ATTRIBUTE);
-
-                    $result3 = $file->setAttribute(self::ATTRIBUTE_PATH_NAME, $codeCloneFile->name());
-                    $this->checkResult($result3, self::ATTRIBUTE_PATH_NAME, self::ACTION_SET_ATTRIBUTE);
+                    $file->setAttribute(self::ATTRIBUTE_LINE_NAME, (string)$codeCloneFile->startLine());
+                    $file->setAttribute(self::ATTRIBUTE_END_LINE_NAME, (string)$codeCloneFile->endLine());
+                    $file->setAttribute(self::ATTRIBUTE_PATH_NAME, $codeCloneFile->name());
                 }
 
                 $codeFragmentEl = $this->document->createElement(self::CODE_FRAGMENT_ELEMENT_NAME);
-                $this->checkResult($codeFragmentEl, self::CODE_FRAGMENT_ELEMENT_NAME, self::ACTION_CREATE_ELEMENT);
-
                 $cdata = $this->document->createCDATASection($this->escapeForXml($clone->lines()));
-                $this->checkResult($cdata, self::CODE_FRAGMENT_ELEMENT_NAME, self::ACTION_CREATE_ELEMENT);
-
-                $result = $codeFragmentEl->appendChild($cdata);
-                $this->checkResult($result, self::CODE_FRAGMENT_ELEMENT_NAME, self::ACTION_APPEND_CHILD);
-
-                $codeFragment = $duplication->appendChild($codeFragmentEl);
-                $this->checkResult($codeFragment, self::CODE_FRAGMENT_ELEMENT_NAME, self::ACTION_APPEND_CHILD);
+                $codeFragmentEl->appendChild($cdata);
+                $duplication->appendChild($codeFragmentEl);
             }
 
             $this->flush();
         } catch (DOMException $exception) {
             throw new LoggerException($exception->getMessage());
         }
-    }
-
-    /**
-     * @throws LoggerException
-     */
-    private function checkResult(DOMNode|false $result, string $name, string $type): void
-    {
-        if (is_object($result)) {
-            return;
-        }
-
-        $error = 'Unknown action: %s';
-
-        if ($type === self::ACTION_CREATE_ELEMENT) {
-            $error = 'Can not create element: %s';
-        } elseif ($type === self::ACTION_APPEND_CHILD) {
-            $error = 'Can not add new child at the end: %s';
-        } elseif ($type === self::ACTION_SET_ATTRIBUTE) {
-            $error = 'Can not set attribute: %s';
-        }
-
-        $this->generateException($error, $name);
-    }
-
-    /**
-     * @throws LoggerException
-     */
-    private function generateException(string $error, string $name): void
-    {
-        throw new LoggerException(sprintf($error, $name));
     }
 }